Remove everything in the repos before installing laravel
After the decision of 3.12.2021 with BPT, SRD is migrating kanff.org to Laravel for more efficiency, routing, building the blog, benefit from the many features of Laravel compared to pure PHP.

Recuperable design and content will be imported again when the laravel app will be ready. Majority of the logic will be rewritten properly in a Laravel way and with a better implementation.